+++ backend/backend_api/models.py
@@ -1,43 +1,1 @@
-from django.db import models
-from django.utils.text import slugify
-from django.urls import reverse
 
-
-class Category(models.Model):
-    name = models.CharField(max_length=100)
-    slug = models.SlugField(unique=True, blank=True)
-
-    def save(self, *args, **kwargs):
-        if not self.slug:
-            self.slug = slugify(self.name)
-        super(Category, self).save(*args, **kwargs)
-
-    def get_absolute_url(self):
-        return reverse('category', args=[self.slug])
-
-    def __str__(self):
-        return self.name
-
-class Product(models.Model):
-    title = models.CharField(max_length=100)
-    menu = models.ForeignKey(Category, on_delete=models.CASCADE)
-    price = models.DecimalField(max_digits=10, decimal_places=2)
-    description = models.TextField()
-    weight = models.DecimalField(max_digits=10, decimal_places=2)
-    energy = models.CharField(max_length=50)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-    visible = models.BooleanField(default=True)
-    is_new = models.BooleanField(default=False)
-
-    def __str__(self):
-        return self.title
-
-
-class ProductImage(models.Model):
-    product = models.ForeignKey(Product, on_delete=models.CASCADE)
-    image_file = models.ImageField(upload_to='images/product_images/')
-    visible = models.BooleanField(default=True)
-
-    def __str__(self):
-        return f"Image for {self.product.title}"
