+++ backend/backend_api/views.py
@@ -1,8 +1,7 @@
 
 from rest_framework import viewsets
 from rest_framework.pagination import PageNumberPagination
-from .models import Category, Product, ProductImage
-from .serializers import CategorySerializer, ProductSerializer, ProductImageSerializer
+
 
 
 class CustomPagination(PageNumberPagination):
